requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
websockets==16.0
edge-tts==7.2.8
google-generativeai==0.8.6
python-multipart==0.0.27